+++ backend/routers/presentations.py
@@ -968,332 +968,3 @@ async def options_presentation_image(presentation_id: int, filename: str):
         }
     )
 
-@router.get("/{presentation_id}/pptx-slides",
-          summary="Get PPTX slides",
-          description="Gets the list of PNG images for a presentation's PPTX slides")
-async def get_pptx_slides(presentation_id: int, db: AsyncSession = Depends(get_db)):
-    """
-    Get the list of PNG images for a presentation's PPTX slides
-    """
-    # Get PPTX step data
-    query = select(PresentationStepModel).where(
-        (PresentationStepModel.presentation_id == presentation_id) &
-        (PresentationStepModel.step == PresentationStep.PPTX.value)
-    )
-    result = await db.execute(query)
-    pptx_step = result.scalars().first()
-    
-    if not pptx_step or pptx_step.status != StepStatus.COMPLETED.value:
-        raise HTTPException(status_code=404, detail="PPTX slides not found or not generated yet")
-    
-    pptx_data = pptx_step.get_result()
-    if not pptx_data or "png_paths" not in pptx_data:
-        raise HTTPException(status_code=404, detail="PPTX slide images not found")
-    
-    # Create image URLs from the paths
-    slide_images = []
-    for i, path in enumerate(pptx_data["png_paths"]):
-        filename = os.path.basename(path)
-        slide_images.append({
-            "slide_index": i,
-            "filename": filename,
-            "url": f"/presentations/{presentation_id}/pptx-slides/{filename}"
-        })
-    
-    return {"slides": slide_images}
-
-@router.get("/{presentation_id}/pptx-slides/{filename}",
-          summary="Get PPTX slide image",
-          description="Gets a specific PPTX slide image by filename")
-async def get_pptx_slide_image(presentation_id: int, filename: str):
-    """
-    Get a specific PPTX slide image by filename
-    """
-    # Construct the path to the image
-    presentation_dir = os.path.join(PRESENTATIONS_STORAGE_DIR, str(presentation_id))
-    image_path = os.path.join(presentation_dir, filename)
-    
-    # Check if the file exists
-    if not os.path.exists(image_path):
-        raise HTTPException(status_code=404, detail="Slide image not found")
-    
-    return FileResponse(image_path)
-
-@router.get("/{presentation_id}/download-pptx",
-          summary="Download PPTX file",
-          description="Downloads the PPTX file for a presentation")
-async def download_pptx(presentation_id: int, filename: str = None, db: AsyncSession = Depends(get_db)):
-    """
-    Download the PPTX file for a presentation
-    """
-    print(f"\n\n=== DOWNLOAD PPTX REQUEST for presentation {presentation_id} ===")
-    if filename:
-        print(f"Requested filename: {filename}")
-    
-    # Get PPTX step data
-    query = select(PresentationStepModel).where(
-        (PresentationStepModel.presentation_id == presentation_id) &
-        (PresentationStepModel.step == PresentationStep.PPTX.value)
-    )
-    result = await db.execute(query)
-    pptx_step = result.scalars().first()
-    
-    if not pptx_step or pptx_step.status != StepStatus.COMPLETED.value:
-        print(f"ERROR: PPTX file not found or not generated yet for presentation {presentation_id}")
-        raise HTTPException(status_code=404, detail="PPTX file not found or not generated yet")
-    
-    pptx_data = pptx_step.get_result()
-    if not pptx_data or "pptx_path" not in pptx_data:
-        print(f"ERROR: PPTX file path not found in result data for presentation {presentation_id}")
-        raise HTTPException(status_code=404, detail="PPTX file not found")
-    
-    print(f"Preparing PPTX download for presentation {presentation_id}")
-    print(f"PPTX path: {pptx_data['pptx_path']}")
-    print(f"File exists: {os.path.exists(pptx_data['pptx_path'])}")
-    print(f"File size: {os.path.getsize(pptx_data['pptx_path']) if os.path.exists(pptx_data['pptx_path']) else 'N/A'} bytes")
-    
-    # Use the specified filename or create a simple one
-    display_filename = filename or f"presentation_{presentation_id}.pptx"
-    if not display_filename.lower().endswith('.pptx'):
-        display_filename += '.pptx'
-    
-    print(f"Setting download filename to: {display_filename}")
-    
-    # Return the PPTX file with extremely explicit download headers
-    headers = {
-        "Content-Disposition": f'attachment; filename="{display_filename}"',
-        "Content-Type": "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-        "X-Content-Type-Options": "nosniff",
-        # Force download in various browsers
-        "Content-Description": "File Transfer",
-        "Content-Transfer-Encoding": "binary",
-        "Expires": "0",
-        "Cache-Control": "must-revalidate, post-check=0, pre-check=0",
-        "Pragma": "public"
-    }
-    
-    response = FileResponse(
-        pptx_data["pptx_path"],
-        media_type="application/vnd.openxmlformats-officedocument.presentationml.presentation",
-        filename=display_filename,
-        headers=headers
-    )
-    
-    print(f"=== DOWNLOAD PPTX RESPONSE PREPARED for presentation {presentation_id} ===\n\n")
-    return response
-
-@router.get("/{presentation_id}/download-pdf",
-          summary="Download PDF file",
-          description="Downloads the PDF version of a presentation")
-async def download_pdf(presentation_id: int, filename: str = None, db: AsyncSession = Depends(get_db)):
-    """
-    Download the PDF version of a presentation
-    """
-    print(f"\n\n=== DOWNLOAD PDF REQUEST for presentation {presentation_id} ===")
-    if filename:
-        print(f"Requested filename: {filename}")
-    
-    # Get PPTX step data
-    query = select(PresentationStepModel).where(
-        (PresentationStepModel.presentation_id == presentation_id) &
-        (PresentationStepModel.step == PresentationStep.PPTX.value)
-    )
-    result = await db.execute(query)
-    pptx_step = result.scalars().first()
-    
-    if not pptx_step or pptx_step.status != StepStatus.COMPLETED.value:
-        print(f"ERROR: PPTX step not completed for presentation {presentation_id}")
-        raise HTTPException(status_code=404, detail="PDF file not found or not generated yet")
-    
-    pptx_data = pptx_step.get_result()
-    if not pptx_data or "pptx_path" not in pptx_data:
-        print(f"ERROR: PPTX file path not found in result data for presentation {presentation_id}")
-        raise HTTPException(status_code=404, detail="Source files not found")
-        
-    # Get the basename of the PPTX file without extension
-    basename = os.path.splitext(os.path.basename(pptx_data["pptx_path"]))[0]
-    pdf_path = os.path.join(os.path.dirname(pptx_data["pptx_path"]), f"{basename}.pdf")
-    
-    print(f"Preparing PDF download for presentation {presentation_id}")
-    print(f"PDF path: {pdf_path}")
-    print(f"File exists: {os.path.exists(pdf_path)}")
-    print(f"File size: {os.path.getsize(pdf_path) if os.path.exists(pdf_path) else 'N/A'} bytes")
-    
-    # Check if PDF exists
-    if not os.path.exists(pdf_path):
-        print(f"ERROR: PDF file not found at {pdf_path}")
-        raise HTTPException(status_code=404, detail="PDF file not found - it may need to be regenerated")
-    
-    # Use the specified filename or create a simple one
-    display_filename = filename or f"presentation_{presentation_id}.pdf"
-    if not display_filename.lower().endswith('.pdf'):
-        display_filename += '.pdf'
-    
-    print(f"Setting download filename to: {display_filename}")
-    
-    # Return the PDF file with extremely explicit download headers
-    headers = {
-        "Content-Disposition": f'attachment; filename="{display_filename}"',
-        "Content-Type": "application/pdf",
-        "X-Content-Type-Options": "nosniff",
-        # Force download in various browsers
-        "Content-Description": "File Transfer",
-        "Content-Transfer-Encoding": "binary",
-        "Expires": "0",
-        "Cache-Control": "must-revalidate, post-check=0, pre-check=0",
-        "Pragma": "public"
-    }
-    
-    response = FileResponse(
-        pdf_path,
-        media_type="application/pdf",
-        filename=display_filename,
-        headers=headers
-    )
-    
-    print(f"=== DOWNLOAD PDF RESPONSE PREPARED for presentation {presentation_id} ===\n\n")
-    return response
-
-@router.get("/slide-types", response_model=SlideTypesResponse,
-          summary="Get available slide types",
-          description="Returns all available slide types and their fields to help with presentation creation")
-async def get_slide_types():
-    """
-    Returns all available slide types and their fields with descriptions.
-    This helps frontend developers understand the available slide structures.
-    """
-    slide_types = [
-        {
-            "type": "welcome",
-            "description": "Opening slide with title and subtitle",
-            "fields": {
-                "title": "Main presentation title",
-                "subtitle": "Optional subtitle or tagline",
-                "author": "Optional author name"
-            }
-        },
-        {
-            "type": "tableofcontents",
-            "description": "Table of contents slide listing the presentation sections",
-            "fields": {
-                "title": "Table of Contents",
-                "sections": "List of section titles (array of strings)",
-            }
-        },
-        {
-            "type": "section",
-            "description": "Section divider slide with a title",
-            "fields": {
-                "title": "Section title"
-            }
-        },
-        {
-            "type": "content",
-            "description": "Standard content slide with title and bulleted content",
-            "fields": {
-                "title": "Slide title",
-                "content": "Bullet points or paragraphs of text"
-            }
-        },
-        {
-            "type": "contentimage",
-            "description": "Content slide with an accompanying image",
-            "fields": {
-                "title": "Slide title",
-                "content": "Bullet points or paragraphs of text",
-                "image": "Reference to the image for this slide"
-            }
-        },
-        {
-            "type": "image",
-            "description": "Full image slide with optional caption",
-            "fields": {
-                "title": "Optional slide title",
-                "caption": "Optional image caption",
-                "image": "Reference to the image for this slide"
-            }
-        },
-        {
-            "type": "quote",
-            "description": "Quote slide with quote text and author",
-            "fields": {
-                "title": "Optional slide title",
-                "quote": "The quote text",
-                "author": "Author of the quote"
-            }
-        },
-        {
-            "type": "conclusion",
-            "description": "Conclusion slide summarizing key points",
-            "fields": {
-                "title": "Conclusion", 
-                "content": "Summary of key points or final thoughts"
-            }
-        },
-        {
-            "type": "thankyou",
-            "description": "Final thank you slide with optional contact information",
-            "fields": {
-                "title": "Thank You",
-                "subtitle": "Optional subtitle or tagline",
-                "contact": "Optional contact information"
-            }
-        }
-    ]
-    
-    return {"slide_types": slide_types} 
-
-@router.post("/{presentation_id}/update-topic", response_model=PresentationResponse,
-            summary="Update presentation topic",
-            description="Updates the presentation topic and reruns research and slide generation")
-async def update_presentation_topic(
-    presentation_id: int,
-    update_data: TopicUpdateRequest,
-    background_tasks: BackgroundTasks,
-    db: AsyncSession = Depends(get_db)
-):
-    try:
-        # Check if presentation exists
-        result = await db.execute(select(Presentation).filter(Presentation.id == presentation_id))
-        presentation = result.scalars().first()
-        
-        if not presentation:
-            raise HTTPException(status_code=404, detail="Presentation not found")
-        
-        # Extract new topic from request data
-        new_topic = update_data.topic
-        
-        # Update the presentation topic in the database
-        presentation.topic = new_topic
-        await db.commit()
-        await db.refresh(presentation)
-        
-        # Start background task to update research and slides
-        background_tasks.add_task(
-            execute_topic_update_task,
-            presentation_id,
-            new_topic,
-            presentation.author
-        )
-        
-        # Return updated presentation
-        return {
-            "id": presentation.id,
-            "name": presentation.name,
-            "topic": new_topic,
-            "author": presentation.author,
-            "created_at": presentation.created_at.isoformat(),
-            "updated_at": presentation.updated_at.isoformat()
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        import traceback
-        error_traceback = traceback.format_exc()
-        print(f"Unexpected error in update_presentation_topic: {str(e)}")
-        print(f"Traceback: {error_traceback}")
-        
-        raise HTTPException(
-            status_code=500,
-            detail=f"An unexpected error occurred: {str(e)}"
-        )
