+++ backend/run_api.py
@@ -9,7 +9,7 @@ load_dotenv()
 if __name__ == "__main__":
     # Run the API server with hot reload in development
     uvicorn.run(
-        "api_new:app", 
+        "api:app",
         host="0.0.0.0", 
         port=8000, 
         reload=True,
