+++ backend/tests/integration/test_api.py
@@ -5,10 +5,10 @@ import sys
 import asyncio
 from fastapi.testclient import TestClient
 
-# Add parent directory to path to import api_new
+# Add parent directory to path to import api
 sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
-from api_new import app
+from api import app
 from database import init_db
 
 # Initialize database for testing
