+++ backend/tests/test_api_structure.py
@@ -6,10 +6,10 @@ from fastapi.testclient import TestClient
 import sys
 import os
 
-# Add parent directory to path to import api_new
+# Add parent directory to path to import api
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from api_new import app
+from api import app
 
 client = TestClient(app)
